Implemented tensorflow model successfully
Wooohoo it works. Now just have to fine tune the thresholds for black and white pixels to make the program run better. It seems '3' is a difficult number for the model to identify.  Models are saved under save_model and the number is their test accuracy.

My model unfortunately revolted whilst I was implementing console line argument input, and decided that '2' was no longer a '2' and instead 'i'. So thanks for that, Mr 85% accuracy.