Updated dataset and trained model
Based on the scope of the project, I do not believe I have to include letters, which SHOULD significantly improve the output of the model when we dont have to predict numbers. Includes operators, which previous dataset did not have, and runs it through CROHME_extractors balance.py to balance the training data with upper bound of 6000, since it overfits otherwise when dataset is too small.

its really not working. please god. please.