Initial implementation of kotlin adaptation to new pipeline
Currently the application hangs I think because it's doing too much work between the image preprocessing and inferences. While inferring it blocks up the main thread and camera becomes unresponsive, throwing a bunch of errors and ultimately forcing the app to quit, which is bad. But at least the preprocessing works, but the model inference as well may take a long time. May have to show progress bar just so user doesn't get demoralised